# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZuSu Trading API",
    "version": "0.1.0",
    "description": "Day-trading automation platform. Phase 1 covers identity, portfolios, the demo broker, the trading gate and the audit log. Endpoints that would place orders do not exist yet — they arrive with the risk engine and order manager."
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zusu_at"
      }
    },
...
  },
  "paths": {
...
    "/api/risk/portfolios/{id}/gate": {
      "get": {
        "summary": "Whether this portfolio may trade right now, and what is blocking it",
        "tags": ["risk"],
        "description": "The same decision the order manager will enforce before submitting an order. Portfolio limits (daily loss, exposure, correlation) join it in Phase 7.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "portfolioId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "blockers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "severity": {
                            "type": "string",
                            "enum": ["BLOCKING", "WARNING"]
                          }
                        },
                        "required": ["code", "message", "severity"],
                        "additionalProperties": false
                      }
                    },
                    "checkedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": ["portfolioId", "allowed", "blockers", "checkedAt"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/portfolios/{id}/limits": {
      "get": {
        "summary": "Active risk limits for a portfolio",
        "tags": ["risk"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "portfolioId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "version": {
                      "type": "integer"
                    },
                    "maxDailyLoss": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxWeeklyLoss": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxPositionSize": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxPortfolioExposurePct": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxSectorExposurePct": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxSymbolExposurePct": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "maxOpenPositions": {
                      "type": "integer"
                    },
                    "maxTradesPerDay": {
                      "type": "integer"
                    },
                    "maxConsecutiveLosses": {
                      "type": "integer"
                    },
                    "maxDrawdownPct": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    }
                  },
                  "required": [
                    "portfolioId",
                    "version",
                    "maxDailyLoss",
                    "maxWeeklyLoss",
                    "maxPositionSize",
                    "maxPortfolioExposurePct",
                    "maxSectorExposurePct",
                    "maxSymbolExposurePct",
                    "maxOpenPositions",
                    "maxTradesPerDay",
                    "maxConsecutiveLosses",
                    "maxDrawdownPct"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/kill-switch": {
      "post": {
        "summary": "Stop all trading",
        "tags": ["risk"],
        "description": "Halts the named portfolio, or every portfolio the caller can reach. Resting entry orders are cancelled; open positions are left untouched and keep being monitored.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  },
                  "portfolioId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "portfoliosHalted": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid"
                      }
                    },
                    "ordersCancelled": {
                      "type": "integer"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "engagedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": ["portfoliosHalted", "ordersCancelled", "reason", "engagedAt"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/portfolios/{id}/resume": {
      "post": {
        "summary": "Release a halt (administrators only)",
        "tags": ["risk"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "portfolioId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tradingState": {
                      "type": "string",
                      "enum": ["ACTIVE", "HALTED", "RECONCILIATION_ERROR"]
                    }
                  },
                  "required": ["portfolioId", "tradingState"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "this deployment"
    }
  ],
...
}